// src/api/generated/report-apis/models/mFAnalysis.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Define Edge Agents API
 * LangGraph-based multi-agent system for financial market analysis
 * OpenAPI spec version: 1.0.0
 */
import type { MFAnalysisReportData } from "./mFAnalysisReportData";

/**
 * Mutual fund analysis.
 */
export interface MFAnalysis {
  /** The id of the mutual fund analysis. */
  id: string;
  /** The scheme code of the mutual fund. */
  scheme_code: number;
  /** The name of the mutual fund scheme. */
  scheme_name: string;
  /** The date of the analysis. */
  date: string;
  data: MFAnalysisReportData;
}